+++ lib/shipit.rb
@@ -30,6 +30,7 @@ require 'faraday-http-cache'
 
 require 'shipit/version'
 
+require 'shipit/octokit_check_runs'
 require 'shipit/flock'
 require 'shipit/github_app'
 require 'shipit/paginator'